Control Barrier Functions

Background

Implementation

ControlBarrierFunctions.ControlBarrierFunction — Type

ControlBarrierFunction

Control barrier function (CBF) defining a safe set as its zero superlevel set.

Fields

h::Function : function defining the safe set h(x) ≥ 0

α::Function : extended class K function a(h(x)) for CBF

∇h::Function : gradient of CBF

Lfh::Function : Lie derivative of CBF along drift vector field f

Lgh::Function : Lie derivative of CBF along control directions g

ControlBarrierFunctions.ControlBarrierFunction — Method

ControlBarrierFunction(h::Function, Σ::ControlAffineSystem, α::Function)

Construct a CBF from a function h, a control affine system Σ, and an extended class K function α.

ControlBarrierFunctions.ControlBarrierFunction — Method

ControlBarrierFunction(h::Function, Σ::ControlAffineSystem)

If no extended class K function provided, default to the identify function.